# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CryptoCampus API
  version: 1.0.0
  description: |
    团队 01 的 HTTP 接口唯一机器可读契约。所有业务接口以 /api/v1 为前缀。
    修改字段、类型、错误码或权限前，必须先评审本文件。
...
servers:
  - url: /api/v1
    description: 当前部署
...
paths:
...
  /drops/text:
    post:
      tags: [Drops]
      operationId: createTextDrop
      summary: 创建文字密信
      x-required-roles: [student, admin, teacher]
      x-jwt: true
      x-sensitive-fields: [content, access_password, session_key, private_key]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTextDropRequest"
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateDropResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /drops/file:
    post:
      tags: [Drops]
      operationId: createFileDrop
      summary: 创建文件密信，最大 100 MiB
      x-required-roles: [student, admin, teacher]
      x-jwt: true
      x-sensitive-fields: [file, access_password, session_key, private_key]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file, ttl_policy, pqc_mode]
              properties:
                file: { type: string, format: binary }
                ttl_policy: { $ref: "#/components/schemas/TtlPolicy" }
                pqc_mode: { type: boolean }
                access_password: { type: string, minLength: 1, maxLength: 128, writeOnly: true }
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateDropResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"
  /drops/{code}:
    get:
      tags: [Drops]
      operationId: getDropMetadata
      summary: 公开获取密信元数据，不返回密文或明文
      x-required-roles: [guest, student, admin, teacher]
      x-jwt: false
      x-sensitive-fields: [ciphertext, plaintext, session_key]
      security: []
      parameters:
        - $ref: "#/components/parameters/DropCode"
      responses:
        "200":
          description: 元数据
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DropMetadata"
        "404":
          $ref: "#/components/responses/NotFound"
  /drops/{code}/extract:
    post:
      tags: [Drops]
      operationId: extractDrop
      summary: 使用提取码和可选口令解密密信
      x-required-roles: [guest, student, admin, teacher]
      x-jwt: false
      x-sensitive-fields: [access_code, access_password, plaintext, session_key, private_key]
      security: []
      parameters:
        - $ref: "#/components/parameters/DropCode"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [access_code]
              properties:
                access_code: { type: string, pattern: "^[A-Z0-9-]{8,32}$", writeOnly: true }
                access_password: { type: string, maxLength: 128, writeOnly: true }
      responses:
        "200":
          description: 文字密信或短内容
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExtractedDrop" }
            application/octet-stream:
              schema: { type: string, format: binary }
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: cc_refresh
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
...
    DropCode:
      name: code
      in: path
      required: true
      schema: { type: string, pattern: "^[A-Za-z0-9]{11,32}$" }
...
  responses:
    Unauthorized:
      description: 未认证、凭据错误或会话失效
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: 资源不存在或为防枚举隐藏存在性
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: 状态冲突、重复消费或幂等冲突
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    RateLimited:
      description: 额度或频率超限
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PayloadTooLarge:
      description: 文件超过当前操作限制；密信文件的课程上限为 100 MiB
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServiceUnavailable:
      description: 密码引擎或 Provider 暂不可用
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id, details]
      properties:
        code: { type: string, pattern: "^[A-Z][A-Z0-9_]+$" }
        message: { type: string }
        request_id: { type: string }
        details: { type: object, additionalProperties: true }
...
    TtlPolicy:
      type: string
      enum: [burn_after_read, hours_24, days_7]
    CreateTextDropRequest:
      type: object
      additionalProperties: false
      required: [content, ttl_policy, pqc_mode]
      properties:
        content: { type: string, minLength: 1, maxLength: 1048576 }
        ttl_policy: { $ref: "#/components/schemas/TtlPolicy" }
        pqc_mode: { type: boolean }
        access_password: { type: string, minLength: 1, maxLength: 128, writeOnly: true }
    CreateDropResponse:
      type: object
      additionalProperties: false
      required: [id, code, access_code, url, expires_at, pqc_mode]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        access_code: { type: string, minLength: 8, maxLength: 32, x-sensitive: true }
        url: { type: string, format: uri-reference }
        expires_at: { type: [string, "null"], format: date-time }
        pqc_mode: { type: boolean }
    DropMetadata:
      type: object
      additionalProperties: false
      required: [code, kind, status, requires_password, burn_after_read, expires_at]
      properties:
        code: { type: string }
        kind: { type: string, enum: [text, file] }
        status: { type: string, enum: [available, consumed, expired, destroyed, cooling_down] }
        requires_password: { type: boolean }
        burn_after_read: { type: boolean }
        filename: { type: [string, "null"] }
        size: { type: [integer, "null"], minimum: 0 }
        expires_at: { type: [string, "null"], format: date-time }
    ExtractedDrop:
      type: object
      additionalProperties: false
      required: [kind, signature_valid, certificate_valid, inspect_record_id]
      properties:
        kind: { type: string, enum: [text, file] }
        content: { type: string, description: 仅文字密信返回 }
        download_url: { type: string, format: uri-reference, description: 仅文件密信返回的短期一次性地址 }
        signature_valid: { type: boolean }
        certificate_valid: { type: boolean }
        inspect_record_id: { type: string, format: uuid }
...
security:
  - bearerAuth: []